+++ backend/src/main/java/com/pickpick/slackevent/application/message/MessageCreatedService.java
@@ -1,14 +1,15 @@
-package com.pickpick.service;
+package com.pickpick.slackevent.application.message;
 
 import com.pickpick.channel.domain.Channel;
 import com.pickpick.channel.domain.ChannelRepository;
-import com.pickpick.controller.dto.SlackMessageDto;
-import com.pickpick.controller.event.SlackEvent;
 import com.pickpick.exception.ChannelNotFoundException;
 import com.pickpick.exception.MemberNotFoundException;
 import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
 import com.pickpick.message.domain.MessageRepository;
+import com.pickpick.slackevent.application.SlackEvent;
+import com.pickpick.slackevent.application.SlackEventService;
+import com.pickpick.slackevent.application.message.dto.SlackMessageDto;
 import java.util.Map;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
