+++ backend/src/main/java/com/pickpick/slackevent/application/message/MessageDeletedService.java
@@ -1,7 +1,8 @@
-package com.pickpick.service;
+package com.pickpick.slackevent.application.message;
 
-import com.pickpick.controller.event.SlackEvent;
 import com.pickpick.message.domain.MessageRepository;
+import com.pickpick.slackevent.application.SlackEvent;
+import com.pickpick.slackevent.application.SlackEventService;
 import java.util.Map;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
