+++ backend/src/test/java/com/pickpick/service/MessageChangedServiceTest.java
@@ -9,6 +9,7 @@ import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
 import com.pickpick.message.domain.Message;
 import com.pickpick.message.domain.MessageRepository;
+import com.pickpick.slackevent.application.message.MessageChangedService;
 import com.pickpick.utils.TimeUtils;
 import java.time.LocalDateTime;
 import java.util.List;
