+++ backend/src/test/java/com/pickpick/service/MessageDeletedServiceTest.java
@@ -8,6 +8,7 @@ import com.pickpick.member.domain.Member;
 import com.pickpick.member.domain.MemberRepository;
 import com.pickpick.message.domain.Message;
 import com.pickpick.message.domain.MessageRepository;
+import com.pickpick.slackevent.application.message.MessageDeletedService;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
